// src/app/api/model/businessDateUpdateRequest.ts
/**
 * Apache Fineract REST API
 *
 * Contact: dev@fineract.apache.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BusinessDateUpdateRequest { 
    date: string;
    dateFormat: string;
    locale: string;
    /**
     * Type of business date
     */
    type: BusinessDateUpdateRequest.TypeEnum;
}
export namespace BusinessDateUpdateRequest {
    export const TypeEnum = {
        BusinessDate: 'BUSINESS_DATE',
        CobDate: 'COB_DATE'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}